Fix release.yml and install.sh for GH releases (#12)
update to build proper zip file for release with updated install script to setup paths and fixed settings.yaml with proper URI

+++ lib/install.sh
@@ -1,7 +1,29 @@
 #!/bin/bash
 mkdir -p ${HOME}/.phylum
-cp -n src/bin/settings.yaml ${HOME}/.phylum/
-cp -f src/bin/phylum-cli.bash ${HOME}/.phylum/
+
+if [ -f src/bin/settings.yaml ]; then
+  cp -f src/bin/settings.yaml ${HOME}/.phylum/
+elif [ -f settings.yaml ]; then
+  cp -f settings.yaml ${HOME}/.phylum/
+else
+  echo "Can't find settings.yaml"
+fi
+
+if [ -f src/bin/phylum-cli.bash ]; then
+  cp -f src/bin/phylum-cli.bash ${HOME}/.phylum/
+elif [ -f phylum-cli.bash ]; then
+  cp -f phylum-cli.bash ${HOME}/.phylum/
+else
+  echo "Can't find phylum-cli.bash"
+fi
+
+if [ -f src/bin/phylum-cli ]; then
+  cp -f src/bin/phylum-cli ${HOME}/.phylum/
+elif [ -f phylum-cli ]; then
+  cp -f phylum-cli ${HOME}/.phylum/
+else
+  echo "Can't find phylum-cli"
+fi
 
 if ! grep -q 'phylum-cli.bash' $HOME/.bashrc ;
 then
